requirements.txt
# Flask Backend
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Firebase
firebase-admin==6.2.0

# WhatsApp Business API
requests==2.31.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.21
flask-sqlalchemy==3.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Dependencies
Werkzeug==2.3.7
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
urllib3==2.0.5
charset-normalizer==3.3.0
idna==3.4
pycparser==2.21
cffi==1.16.0
cryptography==41.0.4
PyJWT==2.8.0
cachecontrol==0.13.1
msgpack==1.0.7 